refactoring views -> views_request
